created internal DebugEffects in CombatReplay
