+++ GW2EIEvtcParser/EIData/CombatReplay/CombatReplay.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser.EIData
 {
@@ -154,6 +155,31 @@ namespace GW2EIEvtcParser.EIData
         {
             PositionPolling(ParserHelper.CombatReplayPollingRate, fightDuration);
             RotationPolling(ParserHelper.CombatReplayPollingRate, fightDuration);
+        }
+
+        internal static void DebugEffects(AbstractSingleActor p, ParsedEvtcLog log, CombatReplay replay, long start = long.MinValue, long end = long.MaxValue)
+        {
+            IReadOnlyList<EffectEvent> tst = log.CombatData.GetEffectEvents();
+            IReadOnlyList<EffectEvent> tst1 = log.CombatData.GetEffectEventsByDst(p.AgentItem);
+            var effectGUIDs1 = tst1.Select(x => log.CombatData.GetEffectGUIDEvent(x.EffectID).ContentGUID).ToList();
+            IReadOnlyList<EffectEvent> tst2 = log.CombatData.GetEffectEvents(p.AgentItem);
+            var effectGUIDs2 = tst2.Select(x => log.CombatData.GetEffectGUIDEvent(x.EffectID).ContentGUID).ToList();
+            foreach (EffectEvent effectEvt in tst)
+            {
+                if (effectEvt.Time <= start || effectEvt.Time >= end)
+                {
+                    continue;
+                }
+                if (effectEvt.IsAroundDst)
+                {
+                    replay.Decorations.Insert(0, new CircleDecoration(true, 0, 180, ((int)effectEvt.Time, (int)effectEvt.Time + 100), "rgba(0, 180, 255, 1.0)", new AgentConnector(log.FindActor(effectEvt.Dst))));
+                }
+                else
+                {
+
+                    replay.Decorations.Insert(0, new CircleDecoration(true, 0, 180, ((int)effectEvt.Time, (int)effectEvt.Time + 100), "rgba(0, 180, 255, 1.0)", new PositionConnector(effectEvt.Position)));
+                }
+            }
         }
     }
 }
